requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.30.0
plotly>=5.15.0
python-dotenv>=1.0.0
